Fix the pushdown
