+++ src/Storages/StorageMerge.cpp
@@ -773,105 +773,6 @@ private:
     QueryTreeNodePtr replacement_table_expression;
 };
 
-bool hasUnknownColumn(const QueryTreeNodePtr & node, QueryTreeNodePtr replacement_table_expression)
-{
-    QueryTreeNodes stack = { node };
-    while (!stack.empty())
-    {
-        auto current = stack.back();
-        stack.pop_back();
-
-        switch (current->getNodeType())
-        {
-            case QueryTreeNodeType::CONSTANT:
-                break;
-            case QueryTreeNodeType::COLUMN:
-            {
-                auto * column_node = current->as<ColumnNode>();
-                auto source = column_node->getColumnSourceOrNull();
-                if (source != replacement_table_expression)
-                    return true;
-                break;
-            }
-            default:
-            {
-                for (const auto & child : current->getChildren())
-                {
-                    if (child)
-                        stack.push_back(child);
-                }
-            }
-        }
-    }
-    return false;
-}
-
-void replaceFilterExpression(
-    QueryTreeNodePtr & expression,
-    const QueryTreeNodePtr & replacement_table_expression,
-    const ContextPtr & context)
-{
-    auto * function = expression->as<FunctionNode>();
-    if (!function)
-        return;
-
-    if (function->getFunctionName() != "and")
-    {
-        if (hasUnknownColumn(expression, replacement_table_expression))
-            expression = nullptr;
-        return;
-    }
-
-    QueryTreeNodes conjunctions;
-    QueryTreeNodes processing{ expression };
-
-    while (!processing.empty())
-    {
-        auto node = std::move(processing.back());
-        processing.pop_back();
-
-        if (auto * function_node = node->as<FunctionNode>())
-        {
-            if (function_node->getFunctionName() == "and")
-                std::copy(
-                    function_node->getArguments().begin(),
-                    function_node->getArguments().end(),
-                    std::back_inserter(processing)
-                );
-            else
-                conjunctions.push_back(node);
-        }
-        else
-        {
-            conjunctions.push_back(node);
-        }
-    }
-
-    std::swap(processing, conjunctions);
-
-    for (const auto & node : processing)
-    {
-        if (!hasUnknownColumn(node, replacement_table_expression))
-            conjunctions.push_back(node);
-    }
-
-    if (conjunctions.empty())
-    {
-        expression = {};
-        return;
-    }
-    if (conjunctions.size() == 1)
-    {
-        expression = conjunctions[0];
-        return;
-    }
-
-    function->getArguments().getNodes() = std::move(conjunctions);
-
-    const auto function_impl = FunctionFactory::instance().get("and", context);
-    function->resolveAsFunction(function_impl->build(function->getArgumentColumns()));
-}
-
 QueryTreeNodePtr replaceTableExpressionAndRemoveJoin(
     QueryTreeNodePtr query,
     QueryTreeNodePtr original_table_expression,
