+++ src/Storages/transformQueryForExternalDatabaseAnalyzer.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <Analyzer/IQueryTreeNode.h>
+#include <Interpreters/Context_fwd.h>
 
 
 namespace DB
 {
 
-ASTPtr getASTForExternalDatabaseFromQueryTree(const QueryTreeNodePtr & query_tree, const QueryTreeNodePtr & table_expression);
+ASTPtr getASTForExternalDatabaseFromQueryTree(ContextPtr context, const QueryTreeNodePtr & query_tree, const QueryTreeNodePtr & table_expression);
 
 }
